+++ src/Components/Definition/JSX.php
@@ -1,0 +1,63 @@
+<?php
+namespace VKBansal\Prism\Components\Definition;
+
+use VKBansal\Prism\Language\AbstractDefinition;
+
+/**
+ * React JSX definition
+ * @package VKBansal\Prism\Definition\JSX
+ * @version 0.1.0
+ * @author Vivek Kumar Bansal <contact@vkbansal.me>
+ * @license MIT
+ */
+class JSX extends AbstractDefinition
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function getName()
+    {
+        return 'jsx';
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function requires()
+    {
+        return ['markup', 'javascript'];
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function definition()
+    {
+        $js = $this->getDefinition('javascript');
+        return $this->extend('markup', $js);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setup()
+    {
+        $tag =& $this->getDefinition('jsx.tag');
+
+        $tag['pattern'] = "/<\/?[\w:-]+\s*(?:\s+[\w:-]+(?:=(?:(\"|')(\\\\?[\w\W])*?\g{1}|[^\s'\">=]+|(\{[\w\W]*?\})))?\s*)*\/?>/i";
+
+        $tag['inside']['attr-value']['pattern'] = "/=[^\{](?:('|\")[\w\W]*?(\g{1})|[^\s>]+)/i";
+
+        $this->insertBefore('jsx.tag.inside', [
+            "script" => [
+                "pattern" => "/=(\{[\w\W]*?\})/i",
+                "inside" => [
+                    "function" => $this->getDefinition("javascript.function"),
+                    "punctuation" => "/[={}[\];(),.:]/",
+                    "keyword" => $this->getDefinition("javascript.keyword")
+                ],
+                "alias" => 'language-javascript'
+            ]
+        ], 'attr-value');
+    }
+}