+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ##master
 - Added `autoload.php` for autolading files without composer
+- Added `React JSX` definition
 
 ##v0.1.0
 - Initial release